refactor: item list 컴포넌트 분리
